+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # Changelog
 
-## [3.3]
+## [3.3.1]
 
 ### - 2024-06-03
 - Activity.class setStateHistory, setStateActivity 함수 추가
     - setStateHistory : 현재 경로를 등록
     - setStateActivity :  현재 클래스를 등록
     - ㅇㅖㅈㅔ ㅈㅏㄱㅅㅓㅇ
+
+- ElementsMangager div,ul,li,textarea,select,button,radio,checkbox,image classList 추가 삭제 기능 및 contains 기능 추가
+- ElementsMangager ElementImage class new add
 
 ### - 2024-06-01
 - v3.2 -> v3.3 업그레이드